basicly quick fixes for project, added permission classes because anyone was able to update or delete or post town, realm, monster, and other fixes.

+++ witcher/contracts/api_views.py
@@ -2,7 +2,7 @@ from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.decorators import api_view, permission_classes, authentication_classes
 from rest_framework.permissions import AllowAny, IsAuthenticated, IsAuthenticatedOrReadOnly
-from rest_framework.authentication import TokenAuthentication
+from rest_framework.authentication import TokenAuthentication, SessionAuthentication
 from django.contrib.auth import get_user_model
 from django.core.paginator import Paginator, EmptyPage
 from django.shortcuts import get_object_or_404
@@ -13,11 +13,15 @@ from django.utils import timezone
 User = get_user_model()
 
 @api_view(['GET', 'POST'])
+@authentication_classes([TokenAuthentication, SessionAuthentication])
+@permission_classes([IsAuthenticatedOrReadOnly])
 def realm_list(request):
     if request.method == 'GET':
         realms = Realm.objects.all()
         serializer = RealmSerializer(realms, many=True)
         return Response(serializer.data)
+    if not request.user.has_perm('contracts.add_realm'):
+        return Response({'detail': 'Brak uprawnień do tworzenia krain.'}, status=status.HTTP_403_FORBIDDEN)
     serializer = RealmSerializer(data=request.data)
     if serializer.is_valid():
         serializer.save()
@@ -25,25 +29,35 @@ def realm_list(request):
     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 @api_view(['GET', 'PUT', 'DELETE'])
+@authentication_classes([TokenAuthentication, SessionAuthentication])
+@permission_classes([IsAuthenticatedOrReadOnly])
 def realm_detail(request, pk):
     realm = Realm.objects.get(pk=pk)
     if request.method == 'GET':
         return Response(RealmSerializer(realm).data)
     elif request.method == 'PUT':
+        if not request.user.has_perm('contracts.change_realm'):
+             return Response(status=status.HTTP_403_FORBIDDEN)
         serializer = RealmSerializer(realm, data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     elif request.method == 'DELETE':
+        if not request.user.has_perm('contracts.delete_realm'):
+             return Response(status=status.HTTP_403_FORBIDDEN)
         realm.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 @api_view(['GET', 'POST'])
+@authentication_classes([TokenAuthentication, SessionAuthentication])
+@permission_classes([IsAuthenticatedOrReadOnly])
 def town_list(request):
     if request.method == 'GET':
         towns = Town.objects.all()
         return Response(TownSerializer(towns, many=True).data)
+    if not request.user.has_perm('contracts.add_town'):
+        return Response({'detail': 'Brak uprawnień do dodawania miast.'}, status=status.HTTP_403_FORBIDDEN)
     serializer = TownSerializer(data=request.data)
     if serializer.is_valid():
         serializer.save()
@@ -51,25 +65,35 @@ def town_list(request):
     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 @api_view(['GET', 'PUT', 'DELETE'])
+@authentication_classes([TokenAuthentication, SessionAuthentication])
+@permission_classes([IsAuthenticatedOrReadOnly])
 def town_detail(request, pk):
     town = get_object_or_404(Town, pk=pk)
     if request.method == 'GET':
         return Response(TownSerializer(town).data)
     elif request.method == 'PUT':
+        if not request.user.has_perm('contracts.change_town'):
+            return Response({'detail': 'Brak uprawnień do edycji miast.'}, status=status.HTTP_403_FORBIDDEN)
         serializer = TownSerializer(town, data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     elif request.method == 'DELETE':
+        if not request.user.has_perm('contracts.delete_town'):
+            return Response({'detail': 'Brak uprawnień do usuwania miast.'}, status=status.HTTP_403_FORBIDDEN)
         town.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 @api_view(['GET', 'POST'])
+@authentication_classes([TokenAuthentication, SessionAuthentication])
+@permission_classes([IsAuthenticatedOrReadOnly])
 def monster_list(request):
     if request.method== 'GET':
         monsters = Monster.objects.all()
         return Response(MonsterSerializer(monsters, many=True).data)
+    if not request.user.has_perm('contracts.add_monster'):
+        return Response({'detail': 'Brak uprawnień do dodawania potworów.'}, status=status.HTTP_403_FORBIDDEN)
     serializer =MonsterSerializer(data=request.data)
     if serializer.is_valid():
         serializer.save()
@@ -77,17 +101,23 @@ def monster_list(request):
     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 @api_view(['GET', 'PUT', 'DELETE'])
+@authentication_classes([TokenAuthentication, SessionAuthentication])
+@permission_classes([IsAuthenticatedOrReadOnly])
 def monster_detail(request, pk):
     monster= Monster.objects.get(pk=pk)
     if request.method== 'GET':
         return Response(MonsterSerializer(monster).data)
     elif request.method== 'PUT':
+        if not request.user.has_perm('contracts.change_monster'):
+            return Response({'detail': 'Brak uprawnień do edycji potworów.'}, status=status.HTTP_403_FORBIDDEN)
         serializer =MonsterSerializer(monster, data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     elif request.method=='DELETE':
+        if not request.user.has_perm('contracts.delete_monster'):
+            return Response({'detail': 'Brak uprawnień do usuwania potworów.'}, status=status.HTTP_403_FORBIDDEN)
         monster.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
@@ -107,7 +137,8 @@ def validate_contract(data):
     return errors
 
 @api_view(['GET','POST'])
-@authentication_classes([TokenAuthentication])
+@authentication_classes([TokenAuthentication, SessionAuthentication])
+@permission_classes([IsAuthenticatedOrReadOnly])
 def contract_list(request):
     if request.method == 'GET':
         qs = Contract.objects.all().order_by('-time_created')
@@ -116,12 +147,7 @@ def contract_list(request):
     errs=validate_contract(data)
     if errs:
         return Response(errs,status=status.HTTP_400_BAD_REQUEST)
-    if request.user.is_authenticated:
-        owner=request.user
-    else:
-        owner=User.objects.first()
-        if not owner:
-            return Response({"error": "Brak usera"}, status=status.HTTP_400_BAD_REQUEST)
+    owner=request.user
     realm=Realm.objects.get(pk=data['realm_id'])
     town=Town.objects.get(pk=data['town_id'])
     monster=Monster.objects.get(pk=data['monster_id'])
@@ -139,13 +165,12 @@ def contract_detail(request, pk):
     return Response(ContractSerializer(obj).data)
 
 @api_view(['PUT'])
-@authentication_classes([TokenAuthentication])
+@authentication_classes([TokenAuthentication,SessionAuthentication])
 @permission_classes([IsAuthenticated])
 def contract_update(request, pk):
-    try:
-        obj=Contract.objects.get(pk=pk)
-    except Contract.DoesNotExist:
-        return Response(status=status.HTTP_404_NOT_FOUND)
+    obj=get_object_or_404(Contract,pk=pk)
+    if not (request.user.is_staff or obj.owner_id == request.user.id):
+        return Response({'detail': 'Forbidden'}, status=status.HTTP_403_FORBIDDEN)
     data=request.data.copy()
     errs=validate_contract(data)
     if errs:
@@ -163,13 +188,12 @@ def contract_update(request, pk):
     return Response(ContractSerializer(obj).data)
 
 @api_view(['DELETE'])
-@authentication_classes([TokenAuthentication])
+@authentication_classes([TokenAuthentication,SessionAuthentication])
 @permission_classes([IsAuthenticated])
 def contract_delete(request, pk):
-    try:
-        obj=Contract.objects.get(pk=pk)
-    except Contract.DoesNotExist:
-        return Response(status=status.HTTP_404_NOT_FOUND)
+    obj = get_object_or_404(Contract, pk=pk)
+    if not (request.user.is_staff or obj.owner_id == request.user.id):
+        return Response({'detail': 'Forbidden'}, status=status.HTTP_403_FORBIDDEN)
     obj.delete()
     return Response(status=status.HTTP_204_NO_CONTENT)
 @api_view(['GET'])
@@ -181,7 +205,7 @@ def contract_search(request):
     return Response(serializer.data)
 
 @api_view(['GET'])
-@authentication_classes([TokenAuthentication])
+@authentication_classes([TokenAuthentication,SessionAuthentication])
 @permission_classes([IsAuthenticated])
 def user_contracts(request):
     qs=(Contract.objects.select_related('realm','town','monster','owner').filter(owner=request.user).order_by('-time_created'))
@@ -197,7 +221,7 @@ def contracts_stats_monthly(request):
     return Response({'year': year, 'items': data})
 
 @api_view(['GET'])
-@authentication_classes([TokenAuthentication])
+@authentication_classes([TokenAuthentication,SessionAuthentication])
 @permission_classes([IsAuthenticated])
 def user_contracts_summary(request):
     qs=Contract.objects.filter(owner=request.user)
